Added tables

+++ src/main/java/com/peter/kist/repository/SemesterRepository.java
@@ -3,5 +3,5 @@ package com.peter.kist.repository;
 import com.peter.kist.model.entity.Semester;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SemesterRepository extends JpaRepository<Semester, Long> {
+public interface SemesterRepository extends JpaRepository<Semester, Integer> {
 }
